feat: Add notification component for displaying announcements
The code changes include adding a new component called Notification, which is responsible for displaying announcements or messages to users. This component is rendered on the client side and can be hidden or shown based on the configuration in the site.config.js file. The component includes functionality to set an expiration time for the notification and allows users to close it. The notification message can be customized through the site.config.js file.